Reference edit dialog now will appear with a minimum size (was collapsed on itself if there were no references). Moved the reference edit classes to a code file under \Windows since it contains winforms specific code. References containing "monogamecontentprocessors" are now ignored.
