+++ Tools/Pipeline/Pipeline.csproj
@@ -59,7 +59,7 @@
     <Compile Include="Common\IProjectItem.cs" />
     <Compile Include="Common\IProjectObserver.cs" />
     <Compile Include="Common\IView.cs" />
-    <Compile Include="Common\PipelineProject.References.cs">
+    <Compile Include="Windows\AssemblyReferenceEditing.cs">
       <SubType>Form</SubType>
     </Compile>
     <Compile Include="Windows\MainView.cs">
